edit parameter type, move test
