+++ types/api/blocks.d.ts
@@ -1,6 +1,7 @@
 import {OutputBlockData, OutputData} from '../data-formats/output-data';
 import {BlockToolData, ToolConfig} from '../tools';
 import {BlockAPI} from './block';
+import {BlockTuneData} from "../block-tunes/block-tune-data";
 
 /**
  * Describes methods to manipulate with Editor`s blocks
@@ -135,9 +136,9 @@ export interface Blocks {
    *
    * @param id - id of the block to update
    * @param data - the new data. Can be partial.
-   * @param tunes
+   * @param tunes - (optional) tune data
    */
-  update(id: string, data: Partial<BlockToolData>, tunes: object): Promise<BlockAPI>;
+  update(id: string, data: Partial<BlockToolData>, tunes?: {[name: string]: BlockTuneData}): Promise<BlockAPI>;
 
   /**
    * Converts block to another type. Both blocks should provide the conversionConfig.
